Add files via upload
This project is the culmination of all the Python-specific material we learned in the data science pre-tech curriculum. Two data sets—one for Spotify's top songs from 2017–2018 and the other for the best-selling phones from 2000–2022—are included in the research. To construct straightforward analyses and graphics of the contents of these datasets, library functions from the matplotlib and pandas packages are used. We chose these datasets for our study because they provided broad information that was applicable to practically all of the world, not just one specific region. People all throughout the world purchase phones and products of various kinds, as well as music. Despite not being in a problem-solution format, the datasets are valuable for analysis and coding. Thank You.